Rename `AppDeveloperTeeTaskRuntimeSecretController` into `AppRuntimeSecretController`
