+++ src/itest/java/com/iexec/sms/ApiClient.java
@@ -27,7 +27,7 @@ public interface ApiClient {
     String API_URL = "http://localhost:${local.server.port}";
 
     @PostMapping("/apps/{appAddress}/secrets/{secretIndex}")
-    ResponseEntity<String> addAppDeveloperAppRuntimeSecret(
+    ResponseEntity<String> addAppRuntimeSecret(
             @RequestHeader("Authorization") String authorization,
             @PathVariable("appAddress") String appAddress,
             @PathVariable("secretIndex") long secretIndex,
@@ -35,7 +35,7 @@ public interface ApiClient {
     );
 
     @RequestMapping(method = RequestMethod.HEAD, path = "/apps/{appAddress}/secrets/{secretIndex}")
-    ResponseEntity<Void> isAppDeveloperAppRuntimeSecretPresent(
+    ResponseEntity<Void> isAppRuntimeSecretPresent(
             @PathVariable String appAddress,
             @PathVariable long secretIndex
     );
