+++ src/itest/java/com/iexec/sms/secret/TeeTaskRuntimeSecretIntegrationTests.java
@@ -66,7 +66,7 @@ public class TeeTaskRuntimeSecretIntegrationTests extends CommonTestSetup {
         addNewSecret(appAddress, secretIndex, secretValue, ownerAddress);
 
         // Check the new secret exists for the API
-        ResponseEntity<Void> secretExistence = apiClient.isAppDeveloperAppRuntimeSecretPresent(appAddress, secretIndex);
+        ResponseEntity<Void> secretExistence = apiClient.isAppRuntimeSecretPresent(appAddress, secretIndex);
         Assertions.assertThat(secretExistence.getStatusCode()).isEqualTo(HttpStatus.NO_CONTENT);
 
         // We check the secret has been added to the database
@@ -97,7 +97,7 @@ public class TeeTaskRuntimeSecretIntegrationTests extends CommonTestSetup {
         // We shouldn't be able to add a new secret to the database with the same appAddress/index
         try {
             final String authorization = getAuthorization(appAddress, secretIndex, secretValue);
-            apiClient.addAppDeveloperAppRuntimeSecret(authorization, appAddress, secretIndex, secretValue);
+            apiClient.addAppRuntimeSecret(authorization, appAddress, secretIndex, secretValue);
             Assertions.fail("A second runtime secret with the same app address and index should be rejected.");
         } catch (FeignException.Conflict ignored) {
             // Having a Conflict exception is what we expect there.
@@ -109,7 +109,7 @@ public class TeeTaskRuntimeSecretIntegrationTests extends CommonTestSetup {
             when(iexecHubService.getOwner(UPPER_CASE_APP_ADDRESS)).thenReturn(ownerAddress);
 
             final String authorization = getAuthorization(UPPER_CASE_APP_ADDRESS, secretIndex, secretValue);
-            apiClient.addAppDeveloperAppRuntimeSecret(authorization, UPPER_CASE_APP_ADDRESS, secretIndex, secretValue);
+            apiClient.addAppRuntimeSecret(authorization, UPPER_CASE_APP_ADDRESS, secretIndex, secretValue);
             Assertions.fail("A second runtime secret with the same index " +
                     "and an app address whose only difference is the case should be rejected.");
         } catch (FeignException.Conflict ignored) {
@@ -129,14 +129,14 @@ public class TeeTaskRuntimeSecretIntegrationTests extends CommonTestSetup {
 
         // At first, no secret should be in the database
         try {
-            apiClient.isAppDeveloperAppRuntimeSecretPresent(appAddress, secretIndex);
+            apiClient.isAppRuntimeSecretPresent(appAddress, secretIndex);
             Assertions.fail("No secret was expected but one has been retrieved.");
         } catch (FeignException.NotFound ignored) {
             // Having a Not Found exception is what we expect there.
         }
 
         // Add a new secret to the database
-        final ResponseEntity<String> secretCreationResult = apiClient.addAppDeveloperAppRuntimeSecret(authorization, appAddress, secretIndex, secretValue);
+        final ResponseEntity<String> secretCreationResult = apiClient.addAppRuntimeSecret(authorization, appAddress, secretIndex, secretValue);
         Assertions.assertThat(secretCreationResult.getStatusCode()).isEqualTo(HttpStatus.NO_CONTENT);
     }
